Initial move of gridding to interpolate
This (rather substantial) commit refactors the current gridding package
to become the interpolate package. This includes making gridding exist as
a wrapper module (that deprecation warns whenever used), pulling out
some of the internals of the old gridding functions to work on a point
basis, and renaming/reorganizing many of the functions to make them fit
better within a more general interpolation package.

This does not yet include moving over interpolation functionality that
currently resides in calc.
