+++ examples/gridding/Point_Interpolation.py
@@ -15,8 +15,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from metpy.cbook import get_test_data
-from metpy.gridding.gridding_functions import (interpolate, remove_nan_observations,
-                                               remove_repeat_coordinates)
+from metpy.interpolate import (interpolate_to_grid, remove_nan_observations,
+                               remove_repeat_coordinates)
 from metpy.plots import add_metpy_logo
 
 
@@ -82,7 +82,7 @@ x, y, temp = remove_repeat_coordinates(x, y, temp)
 ###########################################
 # Scipy.interpolate linear
 # ------------------------
-gx, gy, img = interpolate(x, y, temp, interp_type='linear', hres=75000)
+gx, gy, img = interpolate_to_grid(x, y, temp, interp_type='linear', hres=75000)
 img = np.ma.masked_where(np.isnan(img), img)
 fig, view = basic_map(to_proj)
 mmb = view.pcolormesh(gx, gy, img, cmap=cmap, norm=norm)
@@ -92,7 +92,7 @@ fig.colorbar(mmb, shrink=.4, pad=0, boundaries=levels)
 # Natural neighbor interpolation (MetPy implementation)
 # -----------------------------------------------------
 # `Reference <https://github.com/Unidata/MetPy/files/138653/cwp-657.pdf>`_
-gx, gy, img = interpolate(x, y, temp, interp_type='natural_neighbor', hres=75000)
+gx, gy, img = interpolate_to_grid(x, y, temp, interp_type='natural_neighbor', hres=75000)
 img = np.ma.masked_where(np.isnan(img), img)
 fig, view = basic_map(to_proj)
 mmb = view.pcolormesh(gx, gy, img, cmap=cmap, norm=norm)
@@ -106,8 +106,8 @@ fig.colorbar(mmb, shrink=.4, pad=0, boundaries=levels)
 # grid resolution = 25 km
 #
 # min_neighbors = 1
-gx, gy, img = interpolate(x, y, temp, interp_type='cressman', minimum_neighbors=1, hres=75000,
-                          search_radius=100000)
+gx, gy, img = interpolate_to_grid(x, y, temp, interp_type='cressman', minimum_neighbors=1,
+                                  hres=75000, search_radius=100000)
 img = np.ma.masked_where(np.isnan(img), img)
 fig, view = basic_map(to_proj)
 mmb = view.pcolormesh(gx, gy, img, cmap=cmap, norm=norm)
@@ -119,7 +119,8 @@ fig.colorbar(mmb, shrink=.4, pad=0, boundaries=levels)
 # search_radius = 100km
 #
 # min_neighbors = 3
-gx, gy, img1 = interpolate(x, y, temp, interp_type='barnes', hres=75000, search_radius=100000)
+gx, gy, img1 = interpolate_to_grid(x, y, temp, interp_type='barnes', hres=75000,
+                                   search_radius=100000)
 img1 = np.ma.masked_where(np.isnan(img1), img1)
 fig, view = basic_map(to_proj)
 mmb = view.pcolormesh(gx, gy, img1, cmap=cmap, norm=norm)
@@ -129,8 +130,8 @@ fig.colorbar(mmb, shrink=.4, pad=0, boundaries=levels)
 # Radial basis function interpolation
 # ------------------------------------
 # linear
-gx, gy, img = interpolate(x, y, temp, interp_type='rbf', hres=75000, rbf_func='linear',
-                          rbf_smooth=0)
+gx, gy, img = interpolate_to_grid(x, y, temp, interp_type='rbf', hres=75000, rbf_func='linear',
+                                  rbf_smooth=0)
 img = np.ma.masked_where(np.isnan(img), img)
 fig, view = basic_map(to_proj)
 mmb = view.pcolormesh(gx, gy, img, cmap=cmap, norm=norm)
