+++ examples/gridding/Wind_SLP_Interpolation.py
@@ -17,7 +17,7 @@ import pandas as pd
 
 from metpy.calc import wind_components
 from metpy.cbook import get_test_data
-from metpy.gridding.gridding_functions import interpolate, remove_nan_observations
+from metpy.interpolate import interpolate_to_grid, remove_nan_observations
 from metpy.plots import add_metpy_logo
 from metpy.units import units
 
@@ -43,8 +43,9 @@ x_masked, y_masked, pres = remove_nan_observations(xp, yp, data['slp'].values)
 
 ###########################################
 # Interpolate pressure using Cressman interpolation
-slpgridx, slpgridy, slp = interpolate(x_masked, y_masked, pres, interp_type='cressman',
-                                      minimum_neighbors=1, search_radius=400000, hres=100000)
+slpgridx, slpgridy, slp = interpolate_to_grid(x_masked, y_masked, pres, interp_type='cressman',
+                                              minimum_neighbors=1, search_radius=400000,
+                                              hres=100000)
 
 ##########################################
 # Get wind information and mask where either speed or direction is unavailable
@@ -64,18 +65,18 @@ wind_dir = wind_dir[good_indices]
 # Both will have the same underlying grid so throw away grid returned from v interpolation.
 u, v = wind_components(wind_speed, wind_dir)
 
-windgridx, windgridy, uwind = interpolate(x_masked, y_masked, np.array(u),
-                                          interp_type='cressman', search_radius=400000,
-                                          hres=100000)
+windgridx, windgridy, uwind = interpolate_to_grid(x_masked, y_masked, np.array(u),
+                                                  interp_type='cressman', search_radius=400000,
+                                                  hres=100000)
 
-_, _, vwind = interpolate(x_masked, y_masked, np.array(v), interp_type='cressman',
-                          search_radius=400000, hres=100000)
+_, _, vwind = interpolate_to_grid(x_masked, y_masked, np.array(v), interp_type='cressman',
+                                  search_radius=400000, hres=100000)
 
 ###########################################
 # Get temperature information
 x_masked, y_masked, t = remove_nan_observations(xp, yp, data['temperature'].values)
-tempx, tempy, temp = interpolate(x_masked, y_masked, t, interp_type='cressman',
-                                 minimum_neighbors=3, search_radius=400000, hres=35000)
+tempx, tempy, temp = interpolate_to_grid(x_masked, y_masked, t, interp_type='cressman',
+                                         minimum_neighbors=3, search_radius=400000, hres=35000)
 
 temp = np.ma.masked_where(np.isnan(temp), temp)
