--- metpy/gridding/__init__.py
@@ -1,10 +1,0 @@
-# Copyright (c) 2016 MetPy Developers.
-# Distributed under the terms of the BSD 3-Clause License.
-# SPDX-License-Identifier: BSD-3-Clause
-r"""Provides tools for interpolating irregularly spaced data onto a regular grid."""
-
-from .gridding_functions import *  # noqa: F403
-from .interpolation import *  # noqa: F403
-
-__all__ = gridding_functions.__all__[:]  # pylint: disable=undefined-variable
-__all__.extend(interpolation.__all__)  # pylint: disable=undefined-variable
